add the VecReadWrapper.
The ideal structure of the program would be for EV3FileBuilder
to own the XML reader, but the only way to do that is either own
the vector or the slice with the file contents.

Since owning the slice requires lifetime annotations, then it has to own
the vector allocation. However, for some ungodly reason, Vec<u8> does
not directly implement Read, nor BufRead, so I just made a simple
wrapper which forwards to &[u8]