start parsing method calls
start storing blocks in a hashmap of ids instead of a linked list to do
this. this is because this program is amazing and may have blocks that
are not connected to anything, and therefore would not be stored in the
linked list.

moreover, since links are not stored in the xml, and a table of wires is
stored at the end, there is no point in storing a linked list as in the
end all the links will have to be re-resolved